[FIX] "Never remove from bin" option now works

+++ app/src/main/java/es/agustruiz/anclapp/ui/anchor/BinAnchorActivity.java
@@ -145,17 +145,19 @@ public class BinAnchorActivity extends AppCompatActivity {
         int maxDaysInBin = Integer.parseInt(PreferenceManager.getDefaultSharedPreferences(mContext).getString(
                 mContext.getString(R.string.key_pref_purge_anchors),
                 mContext.getString(R.string.pref_purge_anchors_default_value)));
-        long maxMillisInBin = maxDaysInBin * 24 * 60 * 60 * 1000;
-        long maxTimeLimit = System.currentTimeMillis() - maxMillisInBin;
-        prepareDAO();
-        mAnchorDAO.openReadOnly();
-        List<Anchor> deletedAnchors = mAnchorDAO.getAll(AnchorDAO.QUERY_GET_DELETED);
-        mAnchorDAO.close();
-        for(Anchor anchor:deletedAnchors){
-            if(anchor.getDeletedTimestamp() < maxTimeLimit){
-                mAnchorDAO.openWritable();
-                mAnchorDAO.delete(anchor);
-                mAnchorDAO.close();
+        if(maxDaysInBin>0) {
+            long maxMillisInBin = maxDaysInBin * 24 * 60 * 60 * 1000;
+            long maxTimeLimit = System.currentTimeMillis() - maxMillisInBin;
+            prepareDAO();
+            mAnchorDAO.openReadOnly();
+            List<Anchor> deletedAnchors = mAnchorDAO.getAll(AnchorDAO.QUERY_GET_DELETED);
+            mAnchorDAO.close();
+            for (Anchor anchor : deletedAnchors) {
+                if (anchor.getDeletedTimestamp() < maxTimeLimit) {
+                    mAnchorDAO.openWritable();
+                    mAnchorDAO.delete(anchor);
+                    mAnchorDAO.close();
+                }
             }
         }
     }
